requirements.txt
cloudshell-cli~=5.0
cloudshell-l1-networking-core~=2.0
cloudshell-logging~=2.0